% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coagmet-data.R
\name{get_coagmet_data}
\alias{get_coagmet_data}
\title{Get CoAgMet Data}
\usage{
get_coagmet_data(
  table = c("meta", "daily", "hourly", "five_minute", "soilmoisture_daily",
    "soilmoisture_hourly"),
  stations = NULL,
  geo_region = NULL,
  start_date = NULL,
  end_date = NULL,
  elements = NULL
)
}
\arguments{
\item{table}{A character specifying which data table to get. Can be one of 
\code{"meta", "daily", "hourly", "five_minute", "soilmoisture_daily", "soilmoisture_hourly"}.}

\item{stations}{An optional vector of weather station names to get data for. 
See \url{https://coagmet.colostate.edu/station_index.php}. Defaults to all stations.}

\item{geo_region}{A numeric vector of length 4. Used to specify southwest and northwest
coordinates of the geographic region (box) in which data should be retrieved.}

\item{start_date}{Start date in \code{yyyy-mm-dd} format.}

\item{end_date}{End date in \code{yyyy-mm-dd} format.}

\item{elements}{A character vector of one or more elements found in table. Default is to get all
elements.}
}
\value{
A tibble
}
\description{
Get CoAgMet Data
}
