% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coagmet-data.R
\name{get_coagmet_data}
\alias{get_coagmet_data}
\title{Get CoAgMet Data}
\usage{
get_coagmet_data(
  table = c("meta", "monthly", "daily", "hourly", "five_minute", "soilmoisture_daily",
    "soilmoisture_hourly"),
  source = c("coag", "coop", "snotel", "coco"),
  stations = NULL,
  geo_region = NULL,
  start_date = NULL,
  end_date = NULL,
  elements = NULL
)
}
\arguments{
\item{table}{A character specifying which data table to get. Can be one of 
\code{"meta", "monthly", "daily", "hourly", "five_minute", "soilmoisture_daily", "soilmoisture_hourly"}.
When \code{table = "meta"} source must be either \code{"coag"} or \code{"coop"}.}

\item{source}{A character specifying where to get the data. 
Can be one of \code{"coag", "coop", "snotel", "coco"}.}

\item{stations}{An optional vector of weather station names to get data for. 
See \url{https://coagmet.colostate.edu/station_index.php}.}

\item{geo_region}{A numeric vector of length 4. Used to specify southwest and northwest coordinates
of the geographic region (box) in which data should be retrieved.}

\item{start_date}{Start date in \code{yyyy-mm-dd} format.}

\item{end_date}{End date in \code{yyyy-mm-dd} format.}

\item{elements}{A character vector of one or more elements found in table.}
}
\value{
the results from the search
}
\description{
Get CoAgMet Data
}
